backend/requirements.txt
anthropic>=0.34.0
fastapi>=0.115.0
httpx>=0.27.0
playwright>=1.49.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
supabase>=2.7.4
uvicorn>=0.32.0
